feat: add supplier product info endpoint
